Fail build due to missing kauditd only when SELinux is enabled
With the check added in commit 7566819139e31c95038f9f1c39a2995d1fc93c17,
diskimage-builder fails to build RPM-based images if kauditd is not
running. However, this is only valid for environments where SELinux is
enabled. If SELinux is disabled (which is identified by an empty _runcon
variable), proceed with running setfiles.

Change-Id: I1b056f20a3a55f7333391207d9e1049d25ece041
Closes-Bug: #1779273