Only cap pylint for python2
I noticed that pylint environment doesn't work with Python 3.7; I
guess the pinned version has become incompatible.  From the history
the pin is only for Python 2.7 compatability, so just restrict it to
that (with the basepython for the environment set to python3, it's
questionable if we still support pylint with 2.7, but anyway).

Change-Id: I6a916c8422239a4e1b47dcf28a4569f1637cae7a